Revert #7464 and allow an empty state
Take a new approach to handling a nil remote state payload. Rather than
define an error indicating a nil remote state (which would need to be
moved to an internal package in order to be imported by all packages
which could use it), we create an empty terraform.State with a Serial of
-1 to ensure it's always superseded by any other state. This restores
the previous behavior, while still preventing possible nil values in the
state.

The effect here is the same, because rather than the RefreshState
switching on

     cached != nil && durable == nil

it will switch on

    durable.Serial < cached.Serial

which sets the same value to RefreshResult().